+++ demo/src/components/sections/SrStyleGuideInputs/SrStyleGuideInputs.vue
@@ -46,13 +46,15 @@
     </div>
     <div id="checkbox-inputs" class="sg-block">
       <h4 class="text-h4">Checkbox</h4>
-      <sr-style-guide-item-row class="mb-4" code='<sr-checkbox label="This is a checkbox." />'>
-        <sr-checkbox label="This is a checkbox."/>
+      <sr-style-guide-item-row class="mb-4" code='<sr-checkbox v-model="checkboxValue" label="This is a checkbox." />'>
+        <sr-checkbox label="This is a checkbox." v-model="checkboxValue"/>
       </sr-style-guide-item-row>
-
       <sr-style-guide-item-row code='<sr-checkbox disabled label="This is a disabled checkbox." />'>
         <sr-checkbox disabled label="This is a disabled checkbox."/>
       </sr-style-guide-item-row>
+      <div class="pt-5">
+        Debug checkboxValue: {{ checkboxValue }}
+      </div>
     </div>
     <div id="simple-multi-checkbox-inputs" class="sg-block">
       <h4 class="text-h4">Simple Multi Checkbox</h4>
@@ -169,17 +171,20 @@
     </div>
     <div id="radio-inputs" class="sg-block">
       <h4 class="text-h4">Radio</h4>
-      <sr-style-guide-item-row class="mb-4" code='     <sr-radio-group v-model="radioGroupValue">
-          <sr-radio value="1" label="Option 1"/>
-          <sr-radio value="2" label="Option 2"/>
-          <sr-radio disabled value="3" label="Option 3"/>
-        </sr-radio-group>'>
+      <sr-style-guide-item-row class="mb-4" code='<sr-radio-group v-model="radioGroupValue">
+    <sr-radio value="1" label="Option 1"/>
+    <sr-radio value="2" label="Option 2"/>
+    <sr-radio disabled value="3" label="Option 3"/>
+ </sr-radio-group>'>
         <sr-radio-group v-model="radioGroupValue">
           <sr-radio value="1" label="Option 1"/>
           <sr-radio value="2" label="Option 2"/>
           <sr-radio disabled value="3" label="Option 3"/>
         </sr-radio-group>
       </sr-style-guide-item-row>
+      <p>
+        Debug radioGroupValue: {{ radioGroupValue }}
+      </p>
     </div>
     <div id="date-picker-input" class="sg-block">
       <h4 class="text-h4">Date Picker</h4>
@@ -200,7 +205,7 @@ export default {
             switchValue: true,
             switchValueDense: false,
             radioGroupValue: '1',
-            inputTest: 'asd',
+            checkboxValue: false,
             date: new Date().toISOString().substring(0, 10),
             multiCheckboxOptions: [
                 { label: 'Option 1', value: 1 },