+++ src/components/inputs/SrRadio/SrRadioGroup.vue
@@ -2,6 +2,7 @@
   <v-radio-group
       v-bind="$attrs"
       v-on="$listeners"
+      @change="updateValue"
   >
     <slot></slot>
   </v-radio-group>
@@ -10,6 +11,11 @@
 <script>
 export default {
     name: 'SrRadioGroup',
-    inheritAttrs: false
+    inheritAttrs: false,
+    methods: {
+        updateValue (newValue) {
+            this.$emit('input', newValue)
+        }
+    }
 }
 </script>